src/test/java/org/alan/algorithm/practice/baekjoon/clazz/five/BJ10775Test.java
package org.alan.algorithm.practice.baekjoon.clazz.five;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ10775Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        4
                        3
                        4
                        1
                        1""", "2"),
                Arguments.of("""
                        4
                        6
                        2
                        2
                        3
                        3
                        4
                        4""", "3")
        );
    }

    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ10775().solve(input));
    }
}
